ruff fix and  renames..
