+++ model/modules/feature_transformer/sparse_linear_functions.py
@@ -1,7 +1,6 @@
 import torch
 import torch.nn.functional as F
 from torch import autograd
-import numpy as np
 
 _HAS_CUPY_KERNELS = False
 try:
