+++ tests/test_feature_transformer.py
@@ -6,7 +6,7 @@ import torch
 
 sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-from model.modules.feature_transformer.functions import (
+from model.modules.feature_transformer.sparse_linear_functions import (
     SparseLinearFunction,
 )
